add graphs to account analysis, re-use graphs, delete unused graphs
