+++ src/app/account/account-analysis/account-analysis-results/annual-account-analysis/annual-account-analysis.component.html
@@ -13,7 +13,7 @@
     <app-annual-analysis-summary-table *ngIf="dataDisplay == 'table'" [annualAnalysisSummary]="annualAnalysisSummary"
         [analysisItem]="analysisItem">
     </app-annual-analysis-summary-table>
-    <!-- <app-annual-analysis-summary-graph *ngIf="dataDisplay == 'graph'" [annualAnalysisSummary]="annualAnalysisSummary"
+    <app-annual-analysis-summary-graph *ngIf="dataDisplay == 'graph'" [annualAnalysisSummary]="annualAnalysisSummary"
         [analysisItem]="analysisItem">
-    </app-annual-analysis-summary-graph> -->
+    </app-annual-analysis-summary-graph>
 </div>
