--- src/app/facility/analysis/run-analysis/facility-analysis/annual-facility-analysis/annual-facility-analysis-graph/annual-facility-analysis-graph.component.css
@@ -1,3 +1,0 @@
-g.plot g.trace.bars g.points g.point{
-    opacity: .5 !important;
-}
