--- src/app/facility/analysis/run-analysis/facility-analysis/annual-facility-analysis/annual-facility-analysis-graph/annual-facility-analysis-graph.component.html
@@ -1,15 +1,0 @@
-<!--
-    .facility-bar-chart class references style in assets/styles/plotly.css
-    that sets the opacity of the modeled energy bars with a different opacity.
--->
-<div class="p-2 facility-bar-chart">
-    <div #annualEnergyIntensityAnalysisGraph id="annualEnergyIntensityAnalysisGraph"></div>
-</div>
-<hr class="w-100">
-<div class="pr-2 pl-2">
-    <div #percentImprovementAnalysisGraph id="percentImprovementAnalysisGraph"></div>
-</div>
-<hr class="w-100">
-<div class="pr-2 pl-2">
-    <div #percentAnnualImprovementAnalysisGraph id="percentAnnualImprovementAnalysisGraph"></div>
-</div>
