--- src/app/facility/analysis/run-analysis/group-analysis/monthly-analysis-summary/monthly-analysis-summary-graph/monthly-analysis-summary-graph.component.html
@@ -1,1 +1,0 @@
-<div #monthlyAnalysisGraph id="monthlyAnalysisGraph"></div>
