+++ src/app/shared/shared-analysis/annual-analysis-summary-graph/annual-analysis-summary-graph.component.html
