+++ src/app/shared/shared-analysis/monthly-analysis-summary-graph/monthly-analysis-summary-graph.component.spec.ts
